Extract Log.Entry from Log
